the logic of the sign in controller is transferred to the iterator
